requirements.txt
discord.py
tomli
